requirements.txt
requests
mysql-connector-python
pytz
websocket-client
beautifulsoup4
playwright
